backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
et_xmlfile==2.0.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
pillow==12.1.1
pycparser==3.0
PyJWT==2.12.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
qrcode==8.2
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzdata==2026.1
Werkzeug==3.1.7
